+++ storybook/src/example-pages/Patronen/full-width.ts
@@ -19,14 +19,14 @@ examplePageFactory(
               children: html`
                 <h2>Welkom</h2>
                 <p>Op deze testpagina</p>
-                <h3>What is Lorem Ipsum?</h3>
+                <h3>What is deze pagina?</h3>
                 <p>
-                  Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the
-                  industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and
-                  scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap
-                  into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the
-                  release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing
-                  software like Aldus PageMaker including versions of Lorem Ipsum.
+                  This paragraph is simply dummy text of the printing and typesetting industry. This paragraph has been
+                  the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type
+                  and scrambled it to make a type specimen book. It has survived not only five centuries, but also the
+                  leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s
+                  with the release of Letraset sheets containing this paragraph passages, and more recently with desktop
+                  publishing software like Aldus PageMaker including versions of this paragraph.
                 </p>
               `,
             })}
@@ -68,7 +68,7 @@ examplePageFactory(
               ${anchorTemplate({
                 label: "Timeline of The Hague",
                 url: "https://en.wikipedia.org/wiki/Timeline_of_The_Hague",
-                modifier: "extern",
+                mode: "extern",
               })}
             </p>
             <h3>Early history (h3)</h3>
@@ -96,7 +96,7 @@ examplePageFactory(
               ${anchorTemplate({
                 label: "Wikipedia",
                 url: "https://en.wikipedia.org/wiki/The_Hague",
-                modifier: "extern",
+                mode: "extern",
               })}
             </p>
             <h3>Modern history (h3)</h3>
@@ -126,7 +126,7 @@ examplePageFactory(
               ${anchorTemplate({
                 label: "Wikipedia",
                 url: "https://en.wikipedia.org/wiki/The_Hague",
-                modifier: "extern",
+                mode: "extern",
               })}
             </p>
           `,